Restructure make-command to fix "too many events"
Split make-command into two functions:
 - `onMessage` handles the command lookup and checks
 - `addMessage` ensures that `onMessage` can find a command

As a result of this split, make-command has been renamed to
add-command.
